Extend coin control to implement filter and consolidate functions
1. Changes the (un)select all button to a true select all/select none filp-flop.

2. Implements an input filter function to allow the selection of inputs either less than
or equal to or greater than or equal to a provided input value. This can be used for more
than just consolidations.

3. Implements a "Consolidate" button which limits the selected inputs to 600 to ensure
the consolidation transaction will be successful, presents a pick list to select the destination
address for the consolidation, and then automatically fills in the send transaction for
convenience.