Change Windows/Linux GUI menu bar to a menu button
This removes the standard menu bar from the main application layout in
favor of a menu button placed at the top-left corner of the window. It
enables the visual style of the proposed GUI designs which do not have
a menu bar. The menu button does not appear on macOS because the macOS
standard menu bar at the top of the screen already separates the menus
from the main layout.