Implement GetMaxInputsForConsolidationTxn()
This implements a chain parameter function to get the maximum number of
inputs allowable for a UTXO consolidation transaction with either
the RPC consolidateunspent or the GUI "consolidate" button in coin
control.

The default value for the RPC consolidateunspent function has been
changed to be the same as the upper clamp at the value returned
by GetMaxInputsForConsolidationTxn(). The help returned for
consolidateunspent uses that value as well.