Fix coin control change address status label
The status label for the "custom change address" field in the coin
control section of the "send payment" page was removed by accident
because it overlapped with the name of the change amount label.

This fixes the change address field's status label by changing the
object name so that the page doesn't overwrite the status text for
updates that target the change amount label.