Expand poll registry API to filter by finished polls only
The PollRegistry class provides a sequence iterator that could perform
rudimentary filtering to exclude finished polls, but it could not walk
the sequence for finished polls only. This updates the polls filtering
mechanism to support that use case with a structure that enables other
types of filters for the future.