Fix GUI Inter font rendering on Windows with FreeType
This switches the font rendering engine for Windows to Qt's built-in
FreeType engine and changes the Windows font family from Segoe UI to
Inter. On Windows, Qt rendered the OpenType font poorly--text looked
blurry and distorted. A temporary choice was made to use the system-
provided Segoe UI typeface instead.

Here, we configure the Qt depends recipe to include FreeType support
on Windows and activate that feature when the application starts. We
could switch the format of the Inter font to TrueType, but we'd need
to build it from source since the Inter project doesn't distribute a
desktop TTF set, and the OpenType's "weightier" appearance is easier
to read on a standard-DPI display. FreeType renders non-Latin glyphs
more smoothly on high-DPI desktops as well.