Refresh rainbymagnitude
This fixes the magnitude rounding to agree with the same rounding
method used in magnitude storage in the superblock. It also implements
a dust suppressor by suppressing payments less than one CENT and
renormalizing the remaining payments.

It removes the ability to provide a custom message, because of prior
abuse, and it also implements a minimum requirement of 1000 GRC.