Implement GUI settings for stakesplitting
This commit builds on the implementation of the ArgsManager and implements
GUi settings in a new staking tab in the options dialog to dynamically
control the status of stake splitting for staking optimization, and the
associated settings, the staking efficiency and the minimum post stake split
UTXO value.

Validation is performed on the fields to ensure that the efficiency is between
75% and 98%, and the minstakesplitvalue is 800 or greater.

Note that we are making use of the new updateRwSetting implemented with the
ArgsManager port, which saves runtime settings to a read-write JSON settings file.
Settings in this file take precedence over the read-only settings in the config
file. The tool tips warn the user about this order-of-precedence (override), so
they are not confused later.

This also includes a check-box to disable/enable staking.