Implement dynamic sidestaking entries
This commit introduces two new settings for sidestaking,
-sidestakeaddresses=address1,address2,..., addressN
-sidestakeallocations=alloc1,alloc2,...,allocN

These may be used in both the config file and with changesettings as an alternative to the original
sidestake=address,allocation entries. If the new entries are present, are not empty, and are correctly
filled out then they will take precedence over the older entries. The reason this is necessary is that
the read-write settings file does not support multiple instances of the same setting, so a restructure
was required to support the r-w override.