Refactor to remove the GlobalStatus data structure
This removes the legacy GlobalStatus container which the GUI uses to
poll for information from core state and replaces it with the "push"
pattern employed by other GUI features where we can avoid the locks.
It fixes intermittent block height lag in the GUI during the initial
sync.

Because some information sources for the overview page changed, this
moves the API to fetch the current poll title to the voting model.