Fix fractional scaling for GUI dialog sizes
This provides a workaround to scale the application's GUI dialogs to the
appropriate sizes on desktops with fractional DPI scaling. Under 125% on
Windows, for example, Qt does not adjust the coordinates before versions
5.14 and later so dialogs appear cramped and may obstruct the visibility
of the contents.

This solution applies manual scaling to render the expected dialog sizes
until we can require a minimum Qt version that fully supports fractional
scaling.