Remove use of ArsManager::NETWORK_ONLY attribute
This attribute can't be used yet for Gridcoin, because we still use
separate config files for mainnet/testnet. A comment was made in system.h
to indicate the reserved for future use status of this flag.

This attribute applied to args like addnode was causing them to be ignored
in the testnet config file.