Implement getblocksbatch
Usage:

getblocksbatch <starting block height or hash> <number of blocks> [transaction details]

The return value is JSON object consisting of the number of blocks returned, and the
array of blocks. The blocks will include transaction details if that boolean is set to
true.

Note that the number of blocks must be between 1 and 1000. Also if the starting
height is less than the number of blocks away from the head, the number returned
will be limited to the range between the specified starting height and the head.