Implement PollReference::GetActiveVoteWeight()
This implements the calculation of GetActiveVoteWeight() in a expired or active poll.

To maximize the efficiency of the calculation, the starting and ending block index pointer
for the poll is determined (ending being the current pindexBest if the poll is active),
and a single scan through the index is done.

It has been instrumented with a timer, and based on testing, it takes on average 50 - 1000 msec
depending on the length of the poll.

The calculation is implemented entirely in uint256 arithmetic because it is anticipated that
AVW will be used in certain protocol rules that will be subject to consensus.