Implement close confirmation.
This adds an optional confirmation on close to avoid accidentally closing the
main wallet window. If the user also has enabled minimize on close that will
take priority and the confirmation dialog will not show up until the wallet
is closed from the system tray.

The existing quit handling was slightly alter to more closely resemble that
of the current Bitcoin code base.

Issue: #2211