Update dumpcontracts to allow text dump of minimum info
This commit updates dumpcontracts to allow a dump as a text file of
minimal information to allow external scripting of tests. This mode
is in addition to the original native serialzation dump which is
intended for building the data file for automated unit testing.

It also allows specification of "*" as the contract type to dump all
contracts.