Show GUI error dialog if command line parsing fails
This commit delays the EXIT failure of the application till immediately
after the instantiation of the QApp so that a dialog box informing the
user of a command line parse failure can be raised. This addresses the
effective silent failure of a malformed set of or invalid command line
arguments, since most people that run the GUI app will not see the standard
error output.