Use lambdas in a few connect statements
For actions like overviewAction, SendCoinsAction, etc. we do not
want to pass along the state of QAction::triggered. Only the trigger
signal itself is needed. To accomplish this we use lambdas rather
than methods because they can better be optimized out.

Similar for SignMessageAction and verifyMessageAction, where we
pass in an empty QString.