Scraper enhanced thread safety
This implements qualifiers required for advanced static thread safety
analysis by Clang and other advanced compilers. Note that a few minor
thread safety issues were caught and fixed preparing this commit.

Also, thread safety was implemented around the scraper globals and the
AppCache reads. Even though these are seldom changed and there is little
chance of an issue, they should be completely thread-safe as they
are accessed by multiple threads.

Note that some lambdas were used to do quick locked reads against the
scraper globals in some of the scraper functions to facilitate use of
the global in existing if statements without further change. This is
a little ugly and some of the lambdas are repeated. This will be cleaned
up when the scraper is restructured into classes later.