Implement -rpcwait and -rpcwaittimeout
This is a Bitcoin compatible implementation of -rpcwait and -rpcwaittimeout

Specifying -rpcwait for the rpc client will cause it to wait for the
time specified with -rpcwaittimeout (in seconds) with one second attempts
to connect. If -rpcwaittimeout is specified as zero or less, the wait is
indefinite. Note that the current default value of -rpcwaittimeout is zero.