net: remove obsolete send to IP address command
The reply command which resides in our codebase currently is and has
been unused since the first commit that it has been introduced in. The
complementary `checkorder` and `sendorder` commands has not been
implemented ever, so this command should be fine to remove.

Removing this command allows us to remove the last explicit OpenSSL rand
include.