contract: make reservekey parameter a reference in CreateContractTx
This change ensures that the CReserveKey reserved by
CWallet::CreateTransaction is propagated to the caller of
CreateContractTx. This fixes an issue where the reserved keys are not
kept and returned to the key pool as when CReserveKey is not returned to
the caller, the call to CWallet::CommitTransaction will not be able to
receive and keep the key. This causes change addresses used by contract
transactions(except message contracts which do not use this method but
are specifically crafted) to be used twice.