contract: add Contract::Type::ParseLegacy for parsing old contracts
Per cy's comment on 11aa3a580d32de24dd9b4541ec3377361cc963c4,
Contract::Type::Parse method was designed for use in parsing historical
contracts. Since the aforementioned commit makes changes in the method
which defies this, this commit aims to introduce a new method
LegacyParse which handles the job of parsing pseudo-contracts in the
hashBoinc field and lets the modified Parse method to stay as a method
which strictly de-stringifies the type.