util: clean obsolete config keys
This PR implements a routine which cleans up old config keys from
the config file on startup if the beacon private key is present
in the config file. Since private keys for beacons were moved to
the wallet file in Fern, the existance of these keys can create
a key leak. Since the file is already modified to remove the
private key, other obsolete config keys are also removed.