Correct Record/ForgetMRCRewardBlock
Continue to use pindex for the m_mrc_researchers vector but
use pindex->pprev as the basis for the accrual record, since MRC's
are paid based on the block prior to the staked block. (This is
because when the MRC transaction is initiated, the block prior
to the staked block WAS the head of the chain at that time.