Add block v12 check to createmrcrequest
This adds a check to ensure v12 blocks are enabled at the start
of the createmrcrequest function to avoid a cryptic error message
that the mrc request queue is full if submitted while still on
v11 blocks. That error message is technically true, since the
MRC output limit is zero for v11, but the error message is
misleading.