Modify VerifyTCPPort to use the status of CheckOutboundConnectionCount
Note that the VerifyTCPPort test has historically used the port testing
service portquiz.net. This service has gone down in the past. Since
we are checking outbound and the CheckOutboundConnection test is
done first, if that check did not fail, simply mark the VerifyTCPPort
test successful, since the port must be working properly if successful
outbound connections have been made.