Implement exception handling framework for MRC
This implements a simple exception handling class for MRC that
is an extension of std::runtime_error. It is similar to the bignum_error class.

The MRC_error initializer issues a LogPrintf to the log with the error, so the
throwing function does not have to worry about logging the error separately.

We can get much fancier later, and we need to implement a rigorous exception
handling framework across the board IMHO, but this is good enough for the
intended purpose here, which is to provide more detailed error feedback to
createmrcrequest.