Tweak MRCModel::refresh() to properly handle already submitted MRC
A successfully submitted MRC can be subsequently pushed down in the
queue past the pay limit by other MRCs submitted with higher fees.
This means that the submitted MRC should revert to QUEUE_FULL status.

This also tweaks the WALLET_LOCKED status to only apply if an MRC
is not "found", i.e. there is no submitted MRC present. This is the
correct behavior, because the wallet had to be unlocked to submit
the MRC, and this would represent the situation where the wallet
was relocked after the MRC was submitted. You can't submit a new
MRC until after the existing submitted MRC is either paid or deleted,
so the WALLET_LOCKED status should only apply when there is no MRC
with the walletholder's CPID in the queue.