Fix/optimize ResendWalletTransactions
This commit ensure that only transactions that do not have
a hashBlock (i.e. were in a chain already) are NOT in the main chain
and NOT in the mempool are relayed after additional checks.

The original function did not have the proper filtering, and
was essentially queuing up every transaction in the wallet to be
resent. This bug has been present forever. Note that it was
made worse when the transaction validation was tightened in
5.4.1.0 to strictly check GRC contracts.