Enhance mrc request form
This commit implements functionality to hide the fee boost field
unless the MRC queue is full and fee boost is actually needed to
submit the MRC request. After several months of experience with
MRC in production, it has become apparent that some people were
unnecessarily putting GRC in the fee boost field, or outright
mistaking it for the amount of rewards to redeem.