Use std::filesystem for BackupConfigFile and BackupWallet when Boost ver = 1.74
Note that Boost 1.74 has a known bug that causes copy_file to misbehave
when use across filesystem boundaries (i.e. if the walletbackups
directory is actually a symlink to somewhere else). This commit
is a workaround to use std::filesystem instead for this specific
operation if the wallet is linked against Boost 1.74.

This can be removed when we eventually migrate to std::filesystem
from boost::filesystem.