Resize Header Column with Additional Text
Problem:
The column for "Avg. Credit" in the Researcher Configuration under the Projects Tab is not wide enough to accommodate 1 million+ average credit amounts without showing them in scientific notation. The column is not resizable.

Observations:
It appears that the column width is dictated by the longest field text string or header text length within the column. 
There are 2 exceptions:
- The integer "Avg. Credit" column that is showing a scientific notation instead of resizing to be shown in the standard form.
- The text "CPID" column that automatically resizes when you increase/decrease the width of the overall window.

Solution:
My solution is to add the text "Recent " to the text for the "Avg. Credit" header to accommodate larger numbers in their standard form. This should be sufficient to accommodate the largest RAC in the standard form.

Justification:
I believe this to the easiest and safest way to increase the field length to show larger numbers in their standard form without refactoring the table. 
It will also bring congruence between the terminology used by BOINC and the Gridcoin Wallet.