Split out map for local sidestakes
I realized that a change in design was necessary to split out the map
to hold local sidestakes, because it is possible that a local sidestake
could exist that pays to the same address as a mandatory sidestake.

It is much more correct, and simpler, to retain both instead of coming
up with complicated override rules.