Correct vote display in poll card for walletholder vote(s)
The handling for last vote did not really work correctly in the case
where the walletholder votes multiple times. Because of the way the voting
system handles duplicate outputs in the vote tallying, using the last vote
only does not necessarily contain the entire weight or all the choices for
the walletholder.

This commit implements showing each choice that has weight  and the weight
on each one.