Add missing mrc fees to reward used for SplitCoinStakeOutput
In V13 testing with simultaneous mrcs and mandatory sidestakes,
stakes were being rejected because the calculated total reward
to the staker did not include mrc fees. This apparently has been
an existing omission, but since sidestakes were not validated it
went undetected. Now that mandatory sidestakes are validated in
terms of both destination and amount, they were being rejected
by the validator as being too low.

Adding the mrc fees due to the staker to the staker's reward
fixes the problem.