Change CheckBlockIndexJob to initialization call
This will cause the wallet to ShowChainCorruptedMessage() and
either delete the blockchain files (in the GUI case) and end, or
post a log entry with instructions and end.

A missing pprev on any blockindex entry is fatal and the wallet should
not be allowed to continue operation without rebuilding the chain
data.