Ensure AutoGreylist correctly deals with newly whitelisted projects
This change introduces a map in the projects registry (whitelist)
that tracks when a project was first added. This is used in the
AutoGreylist::RefreshWithSuperblock method.

When projects are newly added to the whitelist, their first entry
(and resultant superblock) will be close to the head of the chain,
well within the 40 superblock lookback to execute the greylist ruleset.

The update loop in RefreshWithSuperblock will only post
greylist entries (remember it goes backwards from the present)
when the entry timestamp is greater than or equal to the first
project entry for that project (i.e. when it was first put on the
whitelist).

This is required to ensure the ZCD and WAS rules work correctly
while newly whitelisted projects do not have the full 40 SBs to
sample.

Also includes some changes due to unit testing issues and locking
optimization